backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
pypdf
python-docx
slowapi
groq
python-dotenv